requirements.txt
PyMuPDF>=1.18.0
Pillow>=9.2.0,<11.0        # Compatible with moviepy 2.1.2
threaded>=4.1.0
numpy>=1.26.0,<2.0         # Compatible with pandas 2.2.0 for Python 3.12
psutil>=6.0.0,<6.1         # Compatible with safety 3.2.9
cachetools>=4.2.0
tqdm>=4.60.0
